Let Spring Boot do the dependency management
Prior to this commit some of the versions/dependencies had a
version which was already managed by Spring Boot or one of
the starters. The version of those dependencies have been
removed and either moved to the <properties> section or
removed all together.

The scope of spring-boot-starter-test has also been fixed
to have <scope>test</scope> to properly scope it. As
spring-boot-starter-test already managed jUnit the explicit
declaration has been removed.

Instead of declaring the JMS dependencies and using a
specific version of ActiveMQ the spring-boot-starter-
activemq has been used to cleanup the pom.

Were possible we removed versions and plugins already
declared in the parent project.