Specifies project.parent.relativePath.
Maven reports warnings if the project, as is, is used as a submodule with a
pom files in the parent directory, because that pom isn't the parent pom.

If we override the default value for the relativePath, maven will look
for the parent pom in the repositories, which is the behavior we want
and eliminates the warnings.

See:
https://maven.apache.org/ref/3.0.5/maven-model/apidocs/org/apache/maven/model/Parent.html#getRelativePath()