Return a UTC ZonedDateTime from CargoRoutingDTO::getArrivalDeadline
By switching the admin/show template from

org.thymeleaf.expression.Dates::format

to

org.thymeleaf.expression.Temporals::format

we avoid an implicit conversion via Date::from(Instant), which
was introducing coupling to user.timezone, introducing an
off-by-one error in the representation of the arrivalDeadline.

Temporals::format is also coupled to user.timezone, therefore
instead of returning an instant we return a ZonedDateTime.  To
ensure that our representation of the output date is consistent with
the parsing of the input date (see CargoAdminController::register),
we specify ZoneOffset.UTC in both places.